Change edition to 2018 and break down into modules
