+++ Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "ostdl"
 version = "0.9.1"
+edition = "2018"
 authors = ["Istvan Szekeres <szekeres@iii.hu>"]
 license = "MIT/Apache-2.0"
 homepage = "https://github.com/Pistahh/ostdl"
